Do not attach CPR information to data constructor ids
because the worker is handled specially by the demand analyser, and the
wrapper is expected to be inlined before that.

There are corner cases (such as undersaturated calls) where this loses
information, but nofib does not know any of these.

On the other side it simplifies and removes code, and it makes it easier
to get holdof the DynFlags whenever we create CPR information.